refactor: `CommentCount` widget follows widget build conventions
